% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDocument.R
\name{createJournalDocument}
\alias{createJournalDocument}
\title{Create the exposure and outcome cohorts}
\usage{
createJournalDocument(resultDirectory, analysisId = 1, cohortIds, cohortNames,
  includeValidation = T)
}
\arguments{
\item{resultDirectory}{The directory containing the results (outputFolder)}

\item{analysisId}{An integer specifying the model Analysis_Id used to create the document}

\item{cohortIds}{A vector of cohort ids}

\item{cohortNames}{A vector of cohort names}

\item{includeValidation}{Whether to look in the validation folder and add any validation results}
}
\description{
Create the exposure and outcome cohorts
}
\details{
This function will create the exposure and outcome cohorts following the definitions included in
this package.
}
