% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlpProtocol.R
\name{createMultiPlpReport}
\alias{createMultiPlpReport}
\title{createMultiPlpReport}
\usage{
createMultiPlpReport(analysisLocation,
  protocolLocation = file.path(analysisLocation, "doc", "protocol.docx"),
  includeModels = F)
}
\arguments{
\item{analysisLocation}{The location of the multiple patient-level prediction study}

\item{protocolLocation}{The location of the auto generated patient-level prediction protocol}

\item{includeModels}{Whether to include the models into the results document}
}
\value{
A work document containing the results of the study is saved into the doc directory in the analysisLocation
}
\description{
Creates a word document report of the prediction
}
\details{
The function creates a word document containing the analysis details, data summary and prediction model results.
}
