% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateShinyApp.R
\name{viewShiny}
\alias{viewShiny}
\title{View shiny app}
\usage{
viewShiny(package = NULL)
}
\arguments{
\item{package}{The name of the package as a string}
}
\description{
View shiny app
}
\details{
This function will open an interactive shiny app for viewing the results
}
\examples{
\dontrun{
viewShiny()
}
}
